ReturnProjectOnGoodBatch search split and passes
The aformentioned unit test has been divided into several individual
tests. They are T_findByBatch_Null, T_findByBatch_Empty,
T_findByBatch_Invalid and T_findByBatch_Valid.  Furthermore, it only
tested to ensure that a valid project was returned and ignored cases
where the batch name provided might not evaluate to an existing batch,
the input is null, or an empty string. All checks for it now pass.

Additionally, a change needed to be made in preTestInt to initialize
dummyList as a new ArrayList<Project>. This prevents a
NullPointerException from *findByBatch* tests.