ShouldReturnLinkedList split and passing
ShouldReturnLinkedList originally tested if a list of projects was
returned from findAllProjects. No validation was being performed on the
size of the list, nor on the possibility of an exception being thrown.

findAllProjects throws a ProjectNotFoundException if the resulting list
is empty. Thus T_findAllProjects_None was created for testing whether
that exceptio is thrown when no projects exist. A second test,
T_findAllProjects_Exist, was created to verify that a list of valid
projects included multiple elements.

Both of these tests pass.