Partial refactor and IfNoDescriptionProvided works
This commit marks a significant attempt at a refactor. To prevent having
to write hundreds of tests for every permutation of three possible
Strings for six or so variables, an attempt has been made to calculate
the permutations for testing.

The current implemenation is not fully complete, but demonstrates
potential. Using JUnit theories, a set of data points can be created
(Strings in this case). A @Theory test can be created which tests each
value in the set of data points. Multiple data point sets can be used to
generate combinations. However, since only two variables hold the
current string for each set, this fails to test every possible because
not all of the variables can have different values. This is still a work
in progress, but I believe I'm on the right track.

testAddProjectIfNoDescriptionIsProvided has been changed to
T_createProjectDTO_NullDescription. It successfully asserts that a
ProjectNotAddedException is thrown by the createProjectDTO.