Adds test for valid findByName and comments
The valid findByName test, T_findByName_Valid, contains two assertions,
that the return type be a list and that the list's size is greater than
0. These two assertions ensure that the datatype is correct and that it
is not empty, which should be the case if a valid name is given.

To help merging with Gerard's branch, I've commented out all tests that I have
not touched.