Successfully implemented console and file logging
The logback-spring.xml contains a logging configuration for
project-service. The console appender is rudimentary and simply
generates output in a particular format. However, FILE is a rolling file
appender. This means that it automatically handles archiving of log
files and deletes them over time. The current defaults are that logs
initially go to ProjectService.log. Each day, that file is moved to an
archive folder, compressed, and renamed to include the date. The maximum
file size before archiving is 10MB. Archived files are kept for a
maximum of 60 days before being automatically purged. Lastly, the
combined total of all log files cannot exceed 20GB.