project-service aspect added
The project-service now contains a basic aspect which advises before,
after, and afterReturning from any method that's an ancestor of the
com.revature package (literally everything). If triggered before a
method, logs the method's name and its paramters.
Triggering after returning from a method logs the method signature, time
of return, and the object returned. Lastly, triggering on an exception
logs the method that threw the exception, exception message, and method
sigature.