Fixed LOG_FILE_NOT_FOUND empty file creation
Spring Boot includes logging that's based on the LogBack logging
framework. It also adds some additional functionality on top.
Originally, I was using logback-spring.xml, which is loaded after some
initialization by Spring. This was causing conflicts, since Spring
expected the properties LOG_FILE and LOG_PATH to either be defined or to
use its default. Using the defaults did not result in the desire I
wanted, so I implemented the functionality I sought in logback.xml,
which is loaded prior to Spring.